# specs/001-restful-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API with JWT Authentication
  description: RESTful API for Todo operations with JWT-protected endpoints
  version: 2.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/v2/users/{user_id}/tasks:
    get:
      summary: Get all tasks for a specific user
      description: Retrieve all tasks belonging to the authenticated user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The user ID to retrieve tasks for
        - in: query
          name: skip
          required: false
          schema:
            type: integer
            default: 0
          description: Number of records to skip
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 100
            maximum: 1000
          description: Maximum number of records to return
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: User not found or no tasks available
    post:
      summary: Create a new task for a user
      description: Create a new task associated with the authenticated user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The user ID to create the task for
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: User not authorized to create tasks for this user
  /api/v2/users/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by ID for the authenticated user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The user ID that owns the task
        - in: path
          name: task_id
          required: true
          schema:
            type: string
          description: The task ID to retrieve
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Task not found
    put:
      summary: Update a specific task
      description: Update all fields of a specific task for the authenticated user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The user ID that owns the task
        - in: path
          name: task_id
          required: true
          schema:
            type: string
          description: The task ID to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Task not found
    delete:
      summary: Delete a specific task
      description: Delete a specific task for the authenticated user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The user ID that owns the task
        - in: path
          name: task_id
          required: true
          schema:
            type: string
          description: The task ID to delete
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Task not found
  /api/v2/users/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status
      description: Toggle the completion status of a specific task for the authenticated user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: The user ID that owns the task
        - in: path
          name: task_id
          required: true
          schema:
            type: string
          description: The task ID to update
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Task not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the task
        title:
          type: string
          description: Task title
          maxLength: 255
        description:
          type: string
          description: Task description
          maxLength: 1000
        completed:
          type: boolean
          description: Whether the task is completed
          default: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Task last update timestamp
    TodoCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Task title
          maxLength: 255
        description:
          type: string
          description: Task description
          maxLength: 1000
        completed:
          type: boolean
          description: Whether the task is completed
          default: false
    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          description: Task title
          maxLength: 255
        description:
          type: string
          description: Task description
          maxLength: 1000
        completed:
          type: boolean
          description: Whether the task is completed
    ApiError:
      type: object
      properties:
        error_code:
          type: string
          description: Standardized error code
        message:
          type: string
          description: Human-readable error description
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Time when the error occurred
  responses:
    UnauthorizedError:
      description: Invalid or missing JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error_code: "UNAUTHORIZED"
            message: "Could not validate credentials"
            timestamp: "2023-10-01T12:00:00Z"
    BadRequestError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error_code: "VALIDATION_ERROR"
            message: "Request validation failed"
            details:
              field: "title"
              reason: "Field required"
            timestamp: "2023-10-01T12:00:00Z"